Split testing utils into external file
